test_sellwin/requirements.txt
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
sqlparse==0.4.3
asgiref==3.6.0
environ==1.0
django_debug_toolbar==3.8.1